requirements.txt
pandas
airtable
pycountry
numpy
python-dotenv
qwikidata
networkx